opentable loading is in testing - not correct

+++ app/src/main/java/com/e/wheretoeat/main/models/Restaurant.kt
@@ -6,7 +6,7 @@ class Restaurant(
     @SerializedName("id") private var restaurantID: Int,
     private var name: String,
     private var address: String,
-    private var price: Int,
+    private var price: Double,
     private var image_url: String
 ) {
